# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CVUploadURLRequest:
    properties:
      file_name:
        example: my-cv.pdf
        type: string
    required:
    - file_name
    type: object
  dto.CVUploadURLResponse:
    properties:
      cv_key:
        example: cvs/da9d5716-4979-4f93-980a-9dfcf00cd02a/my-cv.pdf
        type: string
      upload_url:
        example: https://hireflow-cvs.s3.amazonaws.com/cvs/...?X-Amz-Signature=...
        type: string
    type: object
  dto.CompanyResponse:
    properties:
      created_at:
        example: "2026-05-01T12:00:00Z"
        type: string
      description:
        example: A leading software company.
        type: string
      id:
        example: a1b2c3d4-...
        type: string
      industry:
        example: Technology
        type: string
      location:
        example: Colombo, Sri Lanka
        type: string
      name:
        example: TechCorp Lanka
        type: string
      website:
        example: https://techcorplanka.com
        type: string
    type: object
...
  dto.CreateJobRequest:
    properties:
      description:
        example: We are looking for an experienced Go engineer.
        type: string
      job_type:
        enum:
        - full_time
        - part_time
        - contract
        - internship
        example: full_time
        type: string
      location:
        example: Colombo, Sri Lanka
        maxLength: 150
        type: string
      title:
        example: Senior Go Engineer
        maxLength: 150
        minLength: 2
        type: string
    required:
    - description
    - job_type
    - location
    - title
    type: object
  dto.JobResponse:
    properties:
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      created_at:
        example: "2026-05-01T12:00:00Z"
        type: string
      description:
        example: We are looking for an experienced Go engineer.
        type: string
      id:
        example: b2c3d4e5-...
        type: string
      job_type:
        example: full_time
        type: string
      location:
        example: Colombo, Sri Lanka
        type: string
      status:
        example: open
        type: string
      title:
        example: Senior Go Engineer
        type: string
    type: object
...
  dto.UpdateJobRequest:
    properties:
      description:
        example: Updated description.
        type: string
      job_type:
        enum:
        - full_time
        - part_time
        - contract
        - internship
        example: contract
        type: string
      location:
        example: Remote
        maxLength: 150
        type: string
      title:
        example: Senior Go Engineer (Updated)
        maxLength: 150
        minLength: 2
        type: string
    type: object
...
  handler.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hireflow.com
    name: HireFlow Support
  description: A production-grade Job Recruitment REST API built with Go and Gin.
  license:
    name: MIT
  title: HireFlow API
  version: "1.0"
paths:
...
  /companies/{id}/jobs:
    post:
      consumes:
      - application/json
      description: Creates a job under a company. Recruiter only.
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      - description: Job details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JobResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                errors:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Post a job
      tags:
      - Jobs
  /jobs:
    get:
      description: Returns all open job postings. Public endpoint.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.JobResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: List open jobs
      tags:
      - Jobs
  /jobs/{id}:
    delete:
      description: Deletes a job posting. Only the recruiter who posted it.
      parameters:
      - description: Job UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a job
      tags:
      - Jobs
    get:
      description: Returns a single job posting by ID. Public endpoint.
      parameters:
      - description: Job UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JobResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: Get a job
      tags:
      - Jobs
    put:
      consumes:
      - application/json
      description: Updates a job posting. Only the recruiter who posted it.
      parameters:
      - description: Job UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JobResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a job
      tags:
      - Jobs
...
  /jobs/{id}/close:
    patch:
      description: Marks a job as closed. Only the recruiter who posted it.
      parameters:
      - description: Job UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JobResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Close a job
      tags:
      - Jobs
  /uploads/cv-upload-url:
    post:
      consumes:
      - application/json
      description: Returns a pre-signed S3 URL for uploading a PDF CV directly to
        S3.
      parameters:
      - description: File name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CVUploadURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CVUploadURLResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a CV upload URL
      tags:
      - Uploads
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"